retinal-preprocess-pipeline-2/requirements.txt
opencv-python
tensorflow>=2.11
numpy
pandas
pyyaml
matplotlib
split-folders
Pillow
scikit-learn
